+++ bin/examples/multi_scene_example.py
@@ -9,16 +9,19 @@ WINDOW_WIDTH = 1280
 WINDOW_HEIGHT = 720
 WINDOW_FPS = 60
 
+
 def main():
     # Specify input as a webcam and computer vision model as blazepose
     frame_input = cvgui.Webcam(device_num=0, fps=30)
     cv_model = cvgui.BlazePose()
 
     # Create a pose generator based on a webcam + blazepose
-    pose_input = cvgui.ComputerVisionPose(frame_input=frame_input, model=cv_model)
+    pose_input = cvgui.ComputerVisionPose(
+        frame_input=frame_input, model=cv_model)
 
     # Specify GUI to be pygame
-    ui = cvgui.PyGameUI(width=WINDOW_WIDTH, height=WINDOW_HEIGHT, fps=WINDOW_FPS)
+    ui = cvgui.PyGameUI(width=WINDOW_WIDTH,
+                        height=WINDOW_HEIGHT, fps=WINDOW_FPS)
 
     # Create activity
     activity = cvgui.Activity(pose_input=pose_input, frontend=ui)
@@ -32,7 +35,7 @@ def main():
     # Create a new button
     button_1 = cvgui.button(
         gui=ui,
-        pos=(WINDOW_WIDTH//2,WINDOW_HEIGHT//2),
+        pos=(WINDOW_WIDTH//2, WINDOW_HEIGHT//2),
         activation_distance=50,
         color=(255, 0, 0, 255),
         radius=50
@@ -47,7 +50,7 @@ def main():
 
     button_2 = cvgui.button(
         gui=ui,
-        pos=(WINDOW_WIDTH//2,WINDOW_HEIGHT//2),
+        pos=(WINDOW_WIDTH//2, WINDOW_HEIGHT//2),
         activation_distance=50,
         color=(0, 0, 255, 255),
         radius=50
@@ -82,7 +85,7 @@ def main():
 
     # Create a skeleton to map pose points to
     skeleton = cvgui.skeleton(gui=ui, pos=(WINDOW_WIDTH//2, WINDOW_HEIGHT//2),
-        scale=cv_model.DEFAULT_SCALE)
+                              scale=cv_model.DEFAULT_SCALE)
 
     # Add the skeleton and button to the scene
     scene_1.add_component(button_1)
@@ -94,6 +97,7 @@ def main():
 
     # Start activity
     activity.run()
+
 
 if __name__ == "__main__":
     main()