+++ cvgui/inputs/computer_vision/cv_model/__init__.py
@@ -1,4 +1,2 @@
-"""
-Implemenations of computer vision models that can create pose
-data from a video frame input.
-"""
+"""Implemenations of computer vision models that can create pose \
+data from a video frame input."""