+++ cvgui/outputs/__init__.py
@@ -1,1 +1,4 @@
-from .loggers import *
+"""The `outputs` package contains subpackages \
+related to any output produced by activities."""
+
+from .loggers import *  # noqa